--- include/OLD.db_func.php
@@ -1,54 +1,0 @@
-<?php 
-require("../../protected/authcodes.php");
-
-function db_connect() {
-
-    // Define connection as a static variable, to avoid connecting more than once 
-    static $connection;
-    // Try and connect to the database, if a connection has not been established yet
-    if(!isset($connection)) {
-         // Load configuration as an array. Use the actual location of your configuration file
-        $connection = mysqli_connect(db_host,db_user,db_pass,db_name);
-    }
-
-    // If connection was not successful, handle the error
-    if($connection === false) {
-        // Handle error - notify administrator, log to a file, show an error screen, etc.
-        return mysqli_connect_error(); 
-    }
-    return $connection;
-}
-
-function db_query($query) {
-    // Connect to the database
-    $connection = db_connect();
-
-    // Query the database
-    $result = mysqli_query($connection,$query);
-
-    return $result;
-}
-
-function db_select($query) {
-    $rows = array();
-    $result = db_query($query);
-
-    // If query failed, return `false`
-    if($result === false) {
-        return false;
-    }
-
-    // If query was successful, retrieve all the rows into an array
-    while ($row = mysqli_fetch_assoc($result)) {
-        $rows[] = $row;
-    }
-    return $rows;
-}
-
-function db_error() {
-    $connection = db_connect();
-    return mysqli_error($connection);
-}
-
-
-?>
