+++ include/PHPMailer/load.php
@@ -4,8 +4,6 @@ require('class.phpmailer.php');
 require('class.pop3.php');
 require('class.smtp.php');
 
-//require($_SERVER['DOCUMENT_ROOT'] . "/../protected/authcodes.php"); // grab the server connection details.
-
 //Create a new PHPMailer instance
 function send_user_mail($M_ADDRESS, $M_SUBJECT, $M_MESSAGE) 
 {
