+++ include/db-setup/setup-table_users.php
@@ -1,7 +1,7 @@
 <?php // FUNCTIONS
 
-require($_SERVER['DOCUMENT_ROOT'] . "/../protected/authcodes.php"); // grab the server connection details.
-require($_SERVER['DOCUMENT_ROOT'] . "/template/output.header.php"); // get our output destination ready
+require($_SERVER['DOCUMENT_ROOT'] . "/include/main.php"); // main functions, and references
+require($_SERVER['DOCUMENT_ROOT'] . "/template/output/header.php"); // get our output destination ready
 echo '<pre>'; // prettify my output.part.php stuff
 
 $db = new mysqli(db_host, db_user, db_pass, db_name);
@@ -112,5 +112,5 @@ fail('\'users\' table successfully created.');
 	var_dump($result);
 */
 
-require($_SERVER['DOCUMENT_ROOT'] . "/template/output.footer.php");
+require($_SERVER['DOCUMENT_ROOT'] . "/template/output/footer.php");
 ?>
