+++ user/include/form/register.php
@@ -4,12 +4,8 @@
  * date: Jan 2015
 */
 
-if(!$_SERVER['REQUEST_METHOD'] == 'POST') { exit(); }
-// make sure we're using a form, first thing.
-
-require($_SERVER['DOCUMENT_ROOT'] . "/../protected/authcodes.php"); // grab the server connection details.
-require 'PasswordHash.php'; // for creating the user passwords.
-require($_SERVER['DOCUMENT_ROOT'] . "/template/output.header.php"); // get our output destination ready
+if(!$_SERVER['REQUEST_METHOD'] == 'POST') { exit(); } // make sure we're using a form, first thing.
+require($_SERVER['DOCUMENT_ROOT'] . "/template/output/header.php"); // get our output destination ready
 echo '<pre>'; // prettify my output.part.php stuff
 
 // @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
@@ -167,5 +163,5 @@ if (!$stmt->execute()) {
 
 	}
 // end of code, finish off the theme.
-	require($_SERVER['DOCUMENT_ROOT'] . "/template/output.footer.php");
+	require($_SERVER['DOCUMENT_ROOT'] . "/template/output/footer.php");
 ?>
